Move useEffects and state into parent component
